chore: replace User class with Customer class
